Removed unused files and renamed libraries.py to pdflatex.py as it only deals with generating pdfs
