--- doc/sphinx/_extensions/libraries.py
@@ -1,181 +1,0 @@
-import subprocess
-import string
-import os
-import re
-
-from man import call
-from sphinx.util import logging
-from sphinx.util.console import colorize, term_width_line
-
-# Flag to globally disable generating Doxygen documentation and PDF documentation
-# for libraries as part of the mCRL2 documentation.
-# TODO remove this flag
-_GENERATE_DOXYGEN = False
-_GENERATE_PDF = False
-
-#
-# Some constants and abbreviations
-#
-_LOG = logging.getLogger(__name__)
-_PWD = os.path.abspath(os.path.dirname(__file__))
-_TRUNK = os.environ['MCRL2_SRC_DIR']
-_XML = None
-_RST = None
-_DOXYTAG = None
-_RSTTAG = None
-_PDFTAG = None
-_LIBRARIES = {
-  'atermpp': 'ATerm++',  'bes': 'BES',                      'core': 'Core',
-  'data': 'Data',        'gui': 'GUI',                      'lps': 'LPS',
-  'lts': 'LTS',          'modal_formula': 'Modal Formula',  'smt': 'SMT',
-  'pbes': 'PBES',        'process': 'Process',
-  'trace': 'Trace',      'utilities': 'Utilities'
-}
-_DOXYTEMPLATE = string.Template(
-  open(os.path.join(_PWD, 'libraries_templates', 'doxy.template')).read())
-_REFINDEXTEMPLATE = string.Template(
-  open(os.path.join(_PWD, 'libraries_templates', 'reference.rst.template')).read())
-_ARTICLEINDEXTEMPLATE= string.Template(
-  open(os.path.join(_PWD, 'libraries_templates', 'articles.rst.template')).read())
-
-def get_libraries():
-  """Returns a dictionary of libraries with directoryname as key, and human
-     readable name as value. E.g key='atermpp', value='ATemp++'"""
-  return _LIBRARIES
-
-def log_nonl(msg):
-  _LOG.info(term_width_line(msg), nonl = True)
-
-def log_generated(filename):
-  # log only the last two components of the path (library/file)
-  filename = colorize('darkgreen', str.join('/', filename.split('/')[-2:]))
-  log_nonl(f'transforming Doxygen XML into reST {filename}')
-
-def log_generating_pdf(lib_dir, lib_name):
-  lib_name = colorize('darkgreen', lib_name)
-  lib_dir = colorize('darkgray', f'({lib_dir})')
-  log_nonl(f'compiling LaTeX documents for library {lib_name} {lib_dir}...')
-
-def log_generating_xml(lib_dir, lib_name):
-  lib_name = colorize('darkgreen', lib_name)
-  lib_dir = colorize('darkgray', f'({lib_dir})')
-  log_nonl(f'generating Doxygen XML for library {lib_name} {lib_dir}...')
-
-
-def setvars(temppath):
-  global _XML
-  global _RST
-  global _DOXYTAG
-  global _RSTTAG
-  global _PDFTAG
-  _XML = os.path.join(temppath, 'xml')
-  _RST = os.path.join(_PWD, '..', 'developer_manual')
-  _DOXYTAG = os.path.join(temppath, 'doxy.cached')
-  _RSTTAG = os.path.join(temppath, 'doxyrst.cached')
-  _PDFTAG = os.path.join(temppath, 'libpdf.cached')
-
-
-#
-# Generation functions
-#
-def doxygen(**kwargs):
-  '''Run Doxygen. The configuration used is generated by applying the keyword
-  arguments to this function to _DOXYTEMPLATE.'''
-  call('Doxygen', ['doxygen', '-'],
-       _DOXYTEMPLATE.substitute(kwargs).encode('utf-8'))
-
-
-def xsltproc(src, transform, dst, xmldir):
-  rst = call('xsltproc',
-       ['xsltproc', '--param', 'dir', xmldir, transform, src])
-  open(dst, 'w+').write(rst.decode('utf-8'))
-
-
-def makepdf(src):
-  title = re.search(r'\\title{(.*?)}', open(src + '.tex').read(), re.DOTALL)
-  title = title.group(1) if title else os.path.basename(src)
-  title = ' '.join(title.splitlines())
-  if '{' in title or '\\' in title:
-    title = os.path.basename(src)
-  call('pdflatex', ['pdflatex', src])
-# Only apply bibtex if a "source.bib" file exists in the directory.
-  if os.path.isfile(src + '.bib'):
-    try:
-      call('bibtex', ['bibtex', src])
-      call('pdflatex', ['pdflatex', src])
-    except RuntimeError:
-      pass
-  call('pdflatex', ['pdflatex', src])
-  return title
-
-
-def generate_library_xml(lib_dir, lib_name):
-  '''Generate Doxygen XML for a single library.'''
-  lib_path = os.path.join(_TRUNK, 'libraries', lib_dir)
-  log_generating_xml(lib_dir, lib_name)
-  xml_path = os.path.join(_XML, 'libraries', lib_dir)
-  if not os.path.exists(xml_path):
-    os.makedirs(xml_path)
-  doxygen(
-    INPUT='{0}/include {0}/source'.format(lib_path),
-    PROJECT_NAME=lib_name,
-    PROJECT_NUMBER='unknown',
-    STRIP_FROM_PATH=lib_path,
-    STRIP_FROM_INC_PATH='{0} {1}'.format(
-      os.path.join(lib_path, 'include'),
-      os.path.join(lib_path, 'source')
-    ),
-    XML_OUTPUT=xml_path
-  )
-
-
-def generate_xml():
-  '''Generate doxygen XML for all libraries.'''
-  for dirname, name in _LIBRARIES.items():
-    generate_library_xml(dirname, name)
-  log_nonl('generated Doxygen XML for all libraries')
-  _LOG.info('')
-
-
-def generate_library_pdf(lib_dir, lib_name):
-  '''Search for LaTeX files in lib_dir/latex, compile them and generate
-  an index file called articles.rst.'''
-  subdir = os.path.join('libraries', lib_dir, 'latex')
-  texdir = os.path.join(_RST, subdir)
-  if os.path.exists(texdir):
-    olddir = os.getcwd()
-    os.chdir(texdir)
-    log_generating_pdf(lib_dir, lib_name)
-    titles = []
-    try:
-      for f in os.listdir('.'):
-        if f.endswith('.tex'):
-          fn = os.path.splitext(f)[0]
-          titles.append(':download:`{0} <{2}/{1}.pdf>`'.format(makepdf(fn), fn, subdir))
-      open(os.path.join('..', 'articles.txt'), 'w+').write(
-        _ARTICLEINDEXTEMPLATE.substitute(ARTICLES='* ' + '\n* '.join(titles))
-      )
-    finally:
-      os.chdir(olddir)
-
-
-def generate_rst(temppath):
-  '''Generate reStructuredText documentation for all libraries.'''
-  setvars(temppath)
-
-  doxychanged = False
-
-  if not _GENERATE_DOXYGEN or os.path.exists(_DOXYTAG):
-    _LOG.info('assuming Doxygen XML is up-to-date')
-  else:
-    open(_DOXYTAG, 'w+').close()
-    doxychanged = True
-    generate_xml()
-  if not _GENERATE_PDF or os.path.exists(_PDFTAG):
-      _LOG.info('assuming generated PDF is up-to-date')
-  else:
-    open(_PDFTAG, 'w+').close()
-    for dirname, name in _LIBRARIES.items():
-      generate_library_pdf(dirname, name)
-    log_nonl('compiled LaTeX documents for all libraries')
-    _LOG.info('')
